GUI update, login update

+++ Desktop/MenuItems.cs
@@ -15,10 +15,10 @@ namespace Desktop
     public partial class MenuItems : Form
     {
         private MySqlConnection connection;
-        private string server = "sql7.freemysqlhosting.net";
+        private string server = "cmpg-223-db.ci6pbvbzz3x3.us-west-1.rds.amazonaws.com";
         private string database = "sql7368973";
-        private string uid = "sql7368973";
-        private string password = "1lFxsKtjXr";
+        private string uid = "admin";
+        private string password = "cmpg22310";
         string connectionstring;
         private Desktop myMainForm;
 
@@ -28,9 +28,9 @@ namespace Desktop
             this.myMainForm = MainForm;
         }
 
-        private void addMenuItem( string menuitem, string menuItemDes, int price)
+        private void addMenuItem(string menuitem, string menuItemDes, double price)
         {
-            string addquery = "INSERT INTO MENU_ITEM (Item_Name,Item_Description,Price) VALUES('" + menuitem + "','" + menuItemDes + "','" + price + "')";
+            string addquery = "INSERT INTO MENU-ITEM (Item_Name,Item_Description,Price) VALUES('" + menuitem + "','" + menuItemDes + "','" + price + "')";
             connection.Open();
             MySqlCommand cmd = new MySqlCommand();
             cmd.CommandText = addquery;
@@ -41,7 +41,7 @@ namespace Desktop
 
         private void deleteMenuItem(int menuID)
         {
-            string deletequery = "DELETE FROM MENU_ITEM WHERE Menu_Item_ID ='" + menuID + "'";
+            string deletequery = "DELETE FROM MENU-ITEM WHERE Menu_Item_ID ='" + menuID + "'";
             connection.Open();
             MySqlCommand cmd = new MySqlCommand();
             cmd.CommandText = deletequery;
@@ -49,7 +49,7 @@ namespace Desktop
             cmd.ExecuteNonQuery();
             connection.Close();
 
-            string delete2query = "DELETE FROM ORDER DETAIL WHERE Menu_Item_ID ='" + menuID + "'";
+            string delete2query = "DELETE FROM ORDER DETAIL WHERE MenuItemID ='" + menuID + "'";
             connection.Open();
             cmd = new MySqlCommand();
             cmd.CommandText = delete2query;
@@ -58,9 +58,9 @@ namespace Desktop
             connection.Close();
         }
 
-        private void updateMenuItem(int menuID, string menuitem, string menuItemDes, int price)
+        private void updateMenuItem(int menuID, string menuitem, string menuItemDes, double price)
         {
-            string updatequery = "UPDATE MENU_ITEM SET Menu_Item_ID = '" + menuID + "',Item_Name='" + menuitem + "',Item_Description='" + menuItemDes + "',Price='" + price + "' ";
+            string updatequery = "UPDATE MENU-ITEM SET Menu_Item_ID = '" + menuID + "',Item_Name='" + menuitem + "',Item_Description='" + menuItemDes + "',Price='" + price + "' ";
             connection.Open();
             MySqlCommand cmd = new MySqlCommand();
             cmd.CommandText = updatequery;
@@ -72,9 +72,73 @@ namespace Desktop
 
         private void MenuItems_Load(object sender, EventArgs e)
         {
-            connectionstring = "SERVE=" + server + ";" + "DATABASE=" + database + ";" + "UID" + uid
-                 + ";" + "PASSWORD" + password + ";";
+            connectionstring = "SERVER=" + server + ";" + "DATABASE=" + database + ";" + "UID=" + uid
+                 + ";" + "PASSWORD=" + password + ";";
             connection = new MySqlConnection(connectionstring);
+
+            lblMenuItemHeading.ForeColor = System.Drawing.Color.White;
+            lblDeleteMenuID.ForeColor = System.Drawing.Color.White;
+            lblMenuItemName.ForeColor = System.Drawing.Color.White;
+            lblMenuItemPrice.ForeColor = System.Drawing.Color.White;
+            lblMenuItemDes.ForeColor = System.Drawing.Color.White;
+            lblMenuIDUP.ForeColor = System.Drawing.Color.White;
+            lblMenuNameUP.ForeColor = System.Drawing.Color.White;
+            lblMenuPriceUP.ForeColor = System.Drawing.Color.White;
+            lblMenuDesUP.ForeColor = System.Drawing.Color.White;
+
+            comboBoxDeleteMenuID.ForeColor = System.Drawing.Color.White;
+            comboBoxMenueIDUP.ForeColor = System.Drawing.Color.White;
+
+            textBoxMenuItemName.ForeColor = System.Drawing.Color.White;
+            textBoxMenuItemDes.ForeColor = System.Drawing.Color.White;
+            textBoxMenuItemPrice.ForeColor = System.Drawing.Color.White;
+            textBoxMenuNameUP.ForeColor = System.Drawing.Color.White;
+            textBoxMenuItemDes.ForeColor = System.Drawing.Color.White;
+            textBoxMenuPriceUP.ForeColor = System.Drawing.Color.White;
+
+
+
+            string hex = "#536878";
+            Color color = System.Drawing.ColorTranslator.FromHtml(hex);
+            this.BackColor = color;
+
+            comboBoxDeleteMenuID.BackColor = color;
+            comboBoxMenueIDUP.BackColor = color;
+
+            textBoxMenuItemName.BackColor = color;
+            textBoxMenuItemDes.BackColor = color;
+            textBoxMenuItemPrice.BackColor = color;
+            textBoxMenuNameUP.BackColor = color;
+            textBoxMenuItemDes.BackColor = color;
+            textBoxMenuPriceUP.BackColor = color;
+
+             hex = "#19262d";
+            color = System.Drawing.ColorTranslator.FromHtml(hex);
+            this.BackColor = color;
+            tabPageAddMenuItem.ForeColor = System.Drawing.Color.White;
+            tabPageDeleteMenuItem.ForeColor = System.Drawing.Color.White;
+            tabPageUPmenuItem.ForeColor = System.Drawing.Color.White;
+            tabPageUPmenuItem.BackColor = color;
+            tabPageDeleteMenuItem.BackColor = color;
+            tabPageAddMenuItem.BackColor = color;
+
+            
+
+            
+
+
+
+
+            hex = "#020b0d";
+            color = System.Drawing.ColorTranslator.FromHtml(hex);
+            btnAddMenu.BackColor = color;
+            btnAddMenu.ForeColor = System.Drawing.Color.White;
+
+            btnDeleteMenu.BackColor = color;
+            btnDeleteMenu.ForeColor = System.Drawing.Color.White;
+
+            btnUpMenu.BackColor = color;
+            btnUpMenu.ForeColor = System.Drawing.Color.White;
         }
 
         private void comboBoxDeleteMenuID_Click(object sender, EventArgs e)
@@ -83,7 +147,7 @@ namespace Desktop
             // When combo box clicked all Menu IDs are put in the items
 
 
-            string query = "SELECT * FROM MENU_ITEM";
+            string query = "SELECT * FROM MENU-ITEM";
             //open connection
             connection.Open();
             //put in comand
@@ -104,7 +168,7 @@ namespace Desktop
         {
             // When combo box clicked all Menu IDs are put in the items
 
-            string query = "SELECT * FROM MENU_ITEM";
+            string query = "SELECT * FROM MENU-ITEM";
             //open connection
             connection.Open();
             //put in comand
@@ -129,43 +193,113 @@ namespace Desktop
             {
                 string name = textBoxMenuItemName.Text;
                 string descrip = textBoxMenuItemDes.Text;
-                
 
-                if (int.TryParse(textBoxMenuItemPrice.Text, out int price))
+
+                if (double.TryParse(textBoxMenuItemPrice.Text, out double price))
                 {
                     addMenuItem(name, descrip, price);
                     MessageBox.Show("Menu Item has been  added");
                 }
                 else
                 {
-                   MessageBox.Show("Price has to be a number");
+                    MessageBox.Show("Price has to be a number");
                     textBoxMenuItemPrice.Focus();
                 }
 
-              
-                
+
+
             }
-            catch 
+            catch
             {
                 MessageBox.Show("Menu Item could not be added");
             }
-           
+
         }
 
         private void button2_Click(object sender, EventArgs e)
         {
-            // delete menu item
-            try 
-            {
-                int id = int.Parse(comboBoxDeleteMenuID.SelectedItem.ToString());
+            string payed = "";
+            string delivered = "";
+            bool contain = false;
+            int orderID = 0;
+            string query = "";
+            MySqlCommand cmd = new MySqlCommand(query, connection);
+            MySqlDataReader dataR = cmd.ExecuteReader();
 
+            int id = int.Parse(comboBoxDeleteMenuID.SelectedItem.ToString());
+            // delete menu item
+            
+                try
+                {
+                    // check of order doesnt xist with this menuID
+                    query = "SELECT * FROM ORDERS DETAIL WHERE Menu_ID ='" + id + "'";
+                    //open connection
+                    connection.Open();
+                    //put in comand
+                    cmd = new MySqlCommand(query, connection);
+                    dataR = cmd.ExecuteReader();
+                    // data reader
+                    while (dataR.Read())
+                    {
+                        if (int.Parse(dataR["Menu_ID"] + "") == id)
+                        {
+                            contain = true;
+                        }
+                        orderID = int.Parse(dataR["Order_ID"] + "");
+                    }
+                    // close data reader
+                    dataR.Close();
+                    // close connection
+                    connection.Close();
+                }
+                catch
+                {
+                    deleteMenuItem(id);
+                }
+                
+
+                
+                // check of order doesnt exist with this menuID
+                 query = "SELECT * FROM ORDER WHERE Order_ID ='"+orderID+"'";
+                //open connection
+                connection.Open();
+                //put in comand
+                 cmd = new MySqlCommand(query, connection);
+                dataR = cmd.ExecuteReader();
+                // data reader
+                while (dataR.Read())
+                {
+                    payed = dataR["Paid"] + "";
+                    delivered = dataR["Status"] + "";
+                }
+                // close data reader
+                dataR.Close();
+                // close connection
+                connection.Close();
+
+
+            if ((payed == "0") && (delivered == "0") && (contain))
+
+            {
+                MessageBox.Show("Error.Menu Item could not be deleted. A order that has not been delivered and payed contains this menu item");
+            }
+            else if ((payed == "0") && (delivered == "1") && (contain))
+            {
+                MessageBox.Show("Error.Menu Item could not be deleted. A order that has not been  payed contains this menu item");
+
+            }
+            else
+            {
                 deleteMenuItem(id);
                 MessageBox.Show("Menu Item has been deleted");
             }
-            catch
-            {
-                MessageBox.Show("Error.Menu Item could not be deleted");
-            }
+
+               
+            
+            
+            
+                
+          
             
         }
 
@@ -180,7 +314,7 @@ namespace Desktop
 
                 
 
-                if (int.TryParse(textBoxMenuPriceUP.Text,out int price ))
+                if (double.TryParse(textBoxMenuPriceUP.Text,out double price ))
                 {
                     updateMenuItem(id, itemName, itemDes, price);
                     MessageBox.Show("Menu Item has  been added");
@@ -203,7 +337,7 @@ namespace Desktop
         {
 
             // Menu info linked to Id  is put into update texboxes and comboboxes so it can be edited
-            string query = "SELECT * FROM MENU_ITEM WHERE Menu_Item_ID = '"+int.Parse(comboBoxMenueIDUP.SelectedItem.ToString())+"'";
+            string query = "SELECT * FROM MENU-ITEM WHERE Menu_Item_ID = '"+int.Parse(comboBoxMenueIDUP.SelectedItem.ToString())+"'";
             //open connection
             connection.Open();
             //put in comand
@@ -226,6 +360,11 @@ namespace Desktop
         private void button4_Click(object sender, EventArgs e)
         {
             myMainForm.switchTo("MainForm");
+        }
+
+        private void lblMenuItemHeading_Click(object sender, EventArgs e)
+        {
+
         }
     }
 }
